replace legacy accessor i/f with Selector API
